Minor: move limit_pushdown tests to be with their pass
