+++ datafusion/core/tests/physical_optimizer/mod.rs
@@ -16,6 +16,5 @@
 // under the License.
 
 mod combine_partial_final_agg;
-mod limit_pushdown;
 mod limited_distinct_aggregation;
 mod test_util;
